test: move TAB-focus and darwin notification-event tests to spec/serial/
The 'focus handling' describe in chromium.spec.ts (w.focus() + sendInputEvent
TAB cycling) and the three darwin 'emits show and close events' Notification
tests both depend on OS-singleton state (window focus, Notification Center).
