+++ spec/api-notification.spec.ts
@@ -283,63 +283,6 @@ describe('Notification module', () => {
     expect(n.toastXml).to.equal('<xml/>');
   });
 
-  ifit(process.platform === 'darwin')('emits show and close events', async () => {
-    const n = new Notification({
-      title: 'test notification',
-      body: 'test body',
-      silent: true
-    });
-    {
-      const e = once(n, 'show');
-      n.show();
-      await e;
-    }
-    {
-      const e = once(n, 'close');
-      n.close();
-      await e;
-    }
-  });
-
-  ifit(process.platform === 'darwin')('emits show and close events with custom id', async () => {
-    const n = new Notification({
-      id: 'test-custom-id',
-      title: 'test notification',
-      body: 'test body',
-      silent: true
-    });
-    {
-      const e = once(n, 'show');
-      n.show();
-      await e;
-    }
-    {
-      const e = once(n, 'close');
-      n.close();
-      await e;
-    }
-  });
-
-  ifit(process.platform === 'darwin')('emits show and close events with custom id and groupId', async () => {
-    const n = new Notification({
-      id: 'E017VKL2N8H|C07RBMNS9EK|1772656675.039',
-      groupId: 'E017VKL2N8H|C07RBMNS9EK',
-      title: 'test notification',
-      body: 'test body',
-      silent: true
-    });
-    {
-      const e = once(n, 'show');
-      n.show();
-      await e;
-    }
-    {
-      const e = once(n, 'close');
-      n.close();
-      await e;
-    }
-  });
-
   ifit(process.platform === 'win32')('can show notification with custom id and groupId', () => {
     const n = new Notification({
       id: 'test-custom-id',
